Update vulkan api

+++ Noz_Playground/src/main.cpp
@@ -4,7 +4,7 @@ int Entry(int argc, char** argv)
 {
 	Noz::Application* app = new Noz::Application(NOZ_OBFUSCATE("NOZ_PLAYGROUND"), 900, 600);
 
-	NOZ_LOG_INFO(NOZ_OBFUSCATE("Num of GPU's: %d"), app->GetVulkanAPI().GetGpuCount());
+	NOZ_LOG_INFO(NOZ_OBFUSCATE("Num of GPU's: %d"), (int)app->GetVulkanAPI().GetGpuCount());
 
 	while (app->Alive())
 	{
